integrate bugsnag
* notify bugsnag and terminate the process when net.Listen() false

* simplify error handling logic when accepting tcp connections

* remove some pointless wrapping of errors when a new device connects

* remove call to bugsnag.Configure()

We don't want the default panic handler that bugsnag.Configure() sets
up. Instead, we want more fine-grained control over when to recover from
panics, what to report to bugsnag and when to terminate the process.

* recover panics in goroutines that handle TCP connections and notify bugsnag

* add stacktrace to errors that do not have one yet before notifying bugsnag

* notify bugsnag of all errors that occur during publishing and are instances of bugsnags error type

* adapt "exception" handler to error handling strategy

Instead of relying on a globally configured bugsnag notifier,
create a new instance for every message. Wrap errors occurring
in that handler (i.e. not errors reported from devices) in the
bugsnag error type, so they will end up in bugsnag with a stack
trace.

* use the bugsnag error type in "ota" handler

* use the bugsnag error type in "ping" handler

* use the bugsnag error type in "sentry" handler

* use the bugsnag error type in "stargate" handler

* use the bugsnag error type in "stations" handler

* reduce cyclomatic complexity of getDeviceOS()

* refactor Handler.connect()