integrate datadog
* add package with app-specific monitoring functions

* add config attribute for statsd address

* initialize monitoring during startup

* remove the counter for "processed messages" from the monitoring package

Due to potentially concurrently executing subscribers there is
currently no reliable way to detect when a message has been
processed fully. The metric is also of questionable use since
some messages from devices are just relayed to control clients,
some or modified and some are only used to update the
device/formation state. Therefore it is probably more useful to
look at the number of messages coming in and going out of the
service.

* record number of device clients as they connect and disconnect

* time the request to liberator for fetching device info

* record number of control clients as they connect and disconnect

* count messages that are received from or sent out to the network